move stat_segy to conftest
